chore: remove orphaned imports after menu extraction
